app/requirements.txt
asgiref==3.6.0
attrs==22.2.0
bleach==6.0.0
boto3==1.26.64
botocore==1.29.64
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==3.2.6
django-cors-headers
django-environ
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
drf-spectacular==0.25.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
ndg-httpsclient==0.5.1
Pillow==9.4.0
psycopg2-binary==2.9.1
pusher==3.3.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
